NOBUG return null for random_guess_score as it is incalculable
there are no shown drop zones and we don't know how many drag items have to be
dragged to the image. We could do the complicated calculation to find the area
of a polygon and compare it to the area of the image but it is an overly
complex a calculation I feel outside the scope of the current project.